Add management functions to redgrid
The functions allow to suspend/resume publishing activities to redgrid,
which can help during deployments or cluster rollouts.

If a new cluster rollout happens and all the nodes keep connecting to
each other through redgrid, you may quickly end up with twice the nodes
you expected in the system, with each of them connecting to all the
other ones.

It is saner to suspend redgrid activity on the nodes to be taken out of
rotation while the new cluster is being set up (and hoping no major
netsplit will mess things up in the old one). These functions allow to
do just that.