Connection failure tolerance
When the redgrid service cannot connect to the redis server, do an
incremental backoff and suspend operations up until the connection is
established.

This results in a less obvious failure mode, but will not take down a
node due to repeated failures following a netsplit or outage external to
the node. We're deciding that failing to connect is better than failing
totally as much as partial failures go.

This adds a dependency on the 'backoff' library, which tracks
exponential backoffs.